Include beginnings of door framework and misc fixes
This commit adds some code for 'doors'- areas that change the current
map when touched by the player. The code is currently unpleasant;
doors are hardcoded into the main loop. There is no technical reason
for this, just lack of time at time of writing, so it should be
fixed soon.

This commit also makes various misc. fixes to code structure.
Redundant repeated 'collide is True' statements are replaced
by a tidier structure that assumes a collision unless
False is returned.

The chosen map is no longer a global variable, but is defined in
the main loop, so that it can be changed during the game.
The 'move character' code is separated better from the
collision-detection code.

Some more comments are added for readability.

Sorry, future-me, for the lack of atomicity. I assume nobody else
is using this.