Improve doors and tie characters to maps
This commit improves the doors' implementation, so that doors are tied
to maps. This means that the doors that display will depend on the
map the player is in, and the doors' position will be described in
the map code instead of the main game loop.

The player is now positioned on the screen relative to the door they
have just passed through, to give the illusion of movement. The doors
are also now taller and thinner, to look more like doors.

Characters are now also tied to maps instead of to the main game loop.
The array of characters is currently imported into each map; it would
be better to abstract this out at some point.