Improve map code
This commit switches the map code round, so that the things
drawn to the screen are based on the coordinates of their rects.
Previously, the rects were derived from the onscreen coordinates;
this meant that they could get out of sync, as the onscreen
coordinates needed to remain integers once divided by the size
of a tile for collision detection, etc-- but the code did not ensure
this. It was written that way around solely because that made it
marginally easier to hack up a layout for testing purposes;
there is no good reason to keep it like that now.

This commit also moves the 'rects_to_walls' code into the common
maps class, so that it does not have to be repeated in each map
file. In the future, this should be moved into a class shared by
all components drawn to the screen, as its logic should be shared
by everything onscreen.